+++ tmp-for-ts-migration/00-original/bench/1
@@ -1,0 +1,62 @@
+name=BLS12_381
+ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f200152d
+precomputedMillerLoop2 1736.856 usec
+precomputedMillerLoop2mixed 3875.43 usec
+all ok
+Capi benchmark
+Fr
+Fr::add 0.056 usec
+Fr::sub 0.03 usec
+Fr::mul 0.104 usec
+Fr::sqr 0.106 usec
+Fr::div 14.347 usec
+Fp
+Fp::add 0.038 usec
+Fp::sub 0.037 usec
+Fp::mul 0.27 usec
+Fp::sqr 0.274 usec
+Fp::div 27.263 usec
+Fp2
+Fp2::add 0.057 usec
+Fp2::sub 0.053 usec
+Fp2::mul 0.894 usec
+Fp2::sqr 0.592 usec
+Fp2::div 26.614 usec
+benchmark
+Fr::setByCSPRNG 16.801 usec
+pairing 5331.55 usec
+millerLoop 2345.72 usec
+finalExp 3049.93 usec
+precomputedMillerLoop 1751.407 usec
+G1::add 5.827 usec
+G1::dbl 3.468 usec
+G1::mul 488.983 usec
+G2::add 14.803 usec
+G2::dbl 6.829 usec
+G2::mul 1151.16 usec
+hashAndMapToG1 660.956 usec
+hashAndMapToG2 1419.105 usec
+G1::isValidOrder 363.44 usec
+G2::isValidOrder 462.538 usec
+Fp
+Fp::add 0.662 usec
+Fp::sub 0.641 usec
+Fp::mul 0.911 usec
+Fp::sqr 0.869 usec
+Fp::inv 25.84 usec
+Fp2
+Fp2::add 2.091 usec
+Fp2::sub 1.676 usec
+Fp2::mul 2.397 usec
+Fp2::sqr 2.029 usec
+Fp2::inv 27.994 usec
+Fr
+Fr::add 0.642 usec
+Fr::sub 0.839 usec
+Fr::mul 0.657 usec
+Fr::sqr 0.801 usec
+Fr::inv 14.056 usec
+GT::add 1.94 usec
+GT::mul 19.517 usec
+GT::sqr 13.426 usec
+GT::inv 2.032 usec
