+++ tmp-for-ts-migration/01-after-split-js/bench/2
@@ -1,0 +1,62 @@
+name=BLS12_381
+ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f200152d
+precomputedMillerLoop2 1812.391 usec
+precomputedMillerLoop2mixed 3794.802 usec
+all ok
+Capi benchmark
+Fr
+Fr::add 0.073 usec
+Fr::sub 0.047 usec
+Fr::mul 0.123 usec
+Fr::sqr 0.125 usec
+Fr::div 15.605 usec
+Fp
+Fp::add 0.057 usec
+Fp::sub 0.055 usec
+Fp::mul 0.283 usec
+Fp::sqr 0.293 usec
+Fp::div 27.744 usec
+Fp2
+Fp2::add 0.087 usec
+Fp2::sub 0.073 usec
+Fp2::mul 0.943 usec
+Fp2::sqr 0.601 usec
+Fp2::div 27.759 usec
+benchmark
+Fr::setByCSPRNG 19.422 usec
+pairing 5431.784 usec
+millerLoop 2323.017 usec
+finalExp 2989.174 usec
+precomputedMillerLoop 1758.746 usec
+G1::add 5.943 usec
+G1::dbl 3.758 usec
+G1::mul 475.269 usec
+G2::add 14.705 usec
+G2::dbl 7.047 usec
+G2::mul 1092.791 usec
+hashAndMapToG1 661.756 usec
+hashAndMapToG2 1430.827 usec
+G1::isValidOrder 362.691 usec
+G2::isValidOrder 421.788 usec
+Fp
+Fp::add 0.721 usec
+Fp::sub 0.634 usec
+Fp::mul 0.987 usec
+Fp::sqr 1.108 usec
+Fp::inv 28.5 usec
+Fp2
+Fp2::add 1.469 usec
+Fp2::sub 1.761 usec
+Fp2::mul 2.686 usec
+Fp2::sqr 2.557 usec
+Fp2::inv 32.553 usec
+Fr
+Fr::add 0.682 usec
+Fr::sub 0.905 usec
+Fr::mul 0.8 usec
+Fr::sqr 0.969 usec
+Fr::inv 16.479 usec
+GT::add 1.957 usec
+GT::mul 21.005 usec
+GT::sqr 14.621 usec
+GT::inv 2.473 usec
