+++ tmp-for-ts-migration/01-after-split-js/bench/3
@@ -1,0 +1,62 @@
+name=BLS12_381
+ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f200152d
+precomputedMillerLoop2 1841.528 usec
+precomputedMillerLoop2mixed 3853.7 usec
+all ok
+Capi benchmark
+Fr
+Fr::add 0.07 usec
+Fr::sub 0.043 usec
+Fr::mul 0.117 usec
+Fr::sqr 0.122 usec
+Fr::div 14.115 usec
+Fp
+Fp::add 0.059 usec
+Fp::sub 0.053 usec
+Fp::mul 0.278 usec
+Fp::sqr 0.294 usec
+Fp::div 27.536 usec
+Fp2
+Fp2::add 0.083 usec
+Fp2::sub 0.069 usec
+Fp2::mul 0.93 usec
+Fp2::sqr 0.608 usec
+Fp2::div 27.228 usec
+benchmark
+Fr::setByCSPRNG 16.482 usec
+pairing 5375.734 usec
+millerLoop 2335.659 usec
+finalExp 3064.797 usec
+precomputedMillerLoop 1764.716 usec
+G1::add 5.952 usec
+G1::dbl 3.704 usec
+G1::mul 495.543 usec
+G2::add 14.793 usec
+G2::dbl 6.822 usec
+G2::mul 992.557 usec
+hashAndMapToG1 661.159 usec
+hashAndMapToG2 1399.718 usec
+G1::isValidOrder 365.325 usec
+G2::isValidOrder 427.07 usec
+Fp
+Fp::add 0.716 usec
+Fp::sub 0.647 usec
+Fp::mul 1.002 usec
+Fp::sqr 1.07 usec
+Fp::inv 27.255 usec
+Fp2
+Fp2::add 1.324 usec
+Fp2::sub 2.072 usec
+Fp2::mul 2.588 usec
+Fp2::sqr 2.408 usec
+Fp2::inv 28.728 usec
+Fr
+Fr::add 0.675 usec
+Fr::sub 0.71 usec
+Fr::mul 0.712 usec
+Fr::sqr 0.888 usec
+Fr::inv 13.762 usec
+GT::add 2.015 usec
+GT::mul 19.333 usec
+GT::sqr 13.71 usec
+GT::inv 2.248 usec
