+++ tmp-for-ts-migration/01-after-split-js/exported-field-names.dump
@@ -1,0 +1,64 @@
+[
+  'BLS12_381',
+  'BN254',
+  'BN381_1',
+  'BN381_2',
+  'BN462',
+  'BN_SNARK1',
+  'EC_PROJ',
+  'Fp',
+  'Fp2',
+  'Fr',
+  'G1',
+  'G2',
+  'GT',
+  'IRTF',
+  'NIST_P192',
+  'NIST_P224',
+  'NIST_P256',
+  'PrecomputedG2',
+  'SECP224K1',
+  'SECP256K1',
+  'SECP384R1',
+  'add',
+  'asciiStrToPtr',
+  'curveType',
+  'dbl',
+  'deserializeHexStrToFp',
+  'deserializeHexStrToFp2',
+  'deserializeHexStrToFr',
+  'deserializeHexStrToG1',
+  'deserializeHexStrToG2',
+  'deserializeHexStrToGT',
+  'div',
+  'finalExp',
+  'free',
+  'fromHexStr',
+  'getBasePointG1',
+  'getRandomValues',
+  'hashAndMapToG1',
+  'hashAndMapToG2',
+  'hashToFr',
+  'init',
+  'inv',
+  'millerLoop',
+  'mod',
+  'mul',
+  'mulVec',
+  'neg',
+  'normalize',
+  'pairing',
+  'pow',
+  'precomputedMillerLoop',
+  'precomputedMillerLoop2',
+  'precomputedMillerLoop2mixed',
+  'ptrToAsciiStr',
+  'setETHserialization',
+  'setMapToMode',
+  'sqr',
+  'sub',
+  'toHex',
+  'toHexStr',
+  'verifyOrderG1',
+  'verifyOrderG2'
+]
