+++ tmp-for-ts-migration/01-after-split-js/exported-fields.diff
@@ -1,0 +1,122 @@
+--- 00-original/exported-fields.dump	2021-11-11 20:47:53.000000000 +0900
++++ 01-after-split-js/exported-fields.dump	2021-11-11 20:51:28.000000000 +0900
+@@ -11,14 +11,50 @@
+   NIST_P192: 105,
+   NIST_P224: 106,
+   NIST_P256: 107,
+-  IRTF: 5,
+-  EC_PROJ: 1024,
+-  init: [AsyncFunction (anonymous)],
+-  curveType: 0,
++  IRTF: undefined,
++  EC_PROJ: undefined,
++  Fr: [class (anonymous) extends Common],
++  deserializeHexStrToFr: [Function (anonymous)],
++  Fp: [class (anonymous) extends Common],
++  deserializeHexStrToFp: [Function (anonymous)],
++  Fp2: [class (anonymous) extends Common],
++  deserializeHexStrToFp2: [Function (anonymous)],
++  G1: [class (anonymous) extends Common],
++  deserializeHexStrToG1: [Function (anonymous)],
++  setETHserialization: [Function (anonymous)],
++  setMapToMode: [Function (anonymous)],
++  verifyOrderG1: [Function (anonymous)],
++  verifyOrderG2: [Function (anonymous)],
++  getBasePointG1: [Function (anonymous)],
++  G2: [class (anonymous) extends Common],
++  deserializeHexStrToG2: [Function (anonymous)],
++  GT: [class (anonymous) extends Common],
++  deserializeHexStrToGT: [Function (anonymous)],
++  PrecomputedG2: [class (anonymous)],
++  neg: [Function (anonymous)],
++  sqr: [Function (anonymous)],
++  inv: [Function (anonymous)],
++  normalize: [Function (anonymous)],
++  add: [Function (anonymous)],
++  sub: [Function (anonymous)],
++  mul: [Function (anonymous)],
++  mulVec: [Function (anonymous)],
++  div: [Function (anonymous)],
++  dbl: [Function (anonymous)],
++  hashToFr: [Function (anonymous)],
++  hashAndMapToG1: [Function (anonymous)],
++  hashAndMapToG2: [Function (anonymous)],
++  pow: [Function (anonymous)],
++  pairing: [Function (anonymous)],
++  millerLoop: [Function (anonymous)],
++  precomputedMillerLoop: [Function (anonymous)],
++  precomputedMillerLoop2: [Function (anonymous)],
++  precomputedMillerLoop2mixed: [Function (anonymous)],
++  finalExp: [Function (anonymous)],
+   getRandomValues: [Function: getRandomValues],
+-  mod: <ref *1> {
+-    cryptoGetRandomValues: [Function: _cryptoGetRandomValues],
+-    ready: Promise { [Circular *1] },
++  mod: {
++    cryptoGetRandomValues: [Function: cryptoGetRandomValues],
++    ready: Promise { [Object] },
+     inspect: [Function (anonymous)],
+     preloadedImages: {},
+     preloadedAudios: {},
+@@ -26,7 +62,7 @@
+     _mclBnMalloc: [Function (anonymous)],
+     _mclBnFree: [Function (anonymous)],
+     _mclBn_getVersion: [Function (anonymous)],
+-    _mclBn_init: [Function: 408],
++    _mclBn_init: [Function (anonymous)],
+     _mclBn_getCurveType: [Function (anonymous)],
+     _mclBn_getOpUnitSize: [Function (anonymous)],
+     _mclBn_getG1ByteSize: [Function (anonymous)],
+@@ -485,46 +521,12 @@
+     mclBnGT_setStr: [Function (anonymous)],
+     mclBnGT_getStr: [Function (anonymous)]
+   },
+-  toHex: [Function (anonymous)],
+-  toHexStr: [Function (anonymous)],
++  init: [AsyncFunction (anonymous)],
+   fromHexStr: [Function (anonymous)],
+   free: [Function (anonymous)],
+-  Fr: [class (anonymous) extends Common],
+-  deserializeHexStrToFr: [Function (anonymous)],
+-  Fp: [class (anonymous) extends Common],
+-  deserializeHexStrToFp: [Function (anonymous)],
+-  Fp2: [class (anonymous) extends Common],
+-  deserializeHexStrToFp2: [Function (anonymous)],
+-  G1: [class (anonymous) extends Common],
+-  deserializeHexStrToG1: [Function (anonymous)],
+-  setETHserialization: [Function (anonymous)],
+-  setMapToMode: [Function (anonymous)],
+-  verifyOrderG1: [Function (anonymous)],
+-  verifyOrderG2: [Function (anonymous)],
+-  getBasePointG1: [Function (anonymous)],
+-  G2: [class (anonymous) extends Common],
+-  deserializeHexStrToG2: [Function (anonymous)],
+-  GT: [class (anonymous) extends Common],
+-  deserializeHexStrToGT: [Function (anonymous)],
+-  PrecomputedG2: [class (anonymous)],
+-  neg: [Function (anonymous)],
+-  sqr: [Function (anonymous)],
+-  inv: [Function (anonymous)],
+-  normalize: [Function (anonymous)],
+-  add: [Function (anonymous)],
+-  sub: [Function (anonymous)],
+-  mul: [Function (anonymous)],
+-  mulVec: [Function (anonymous)],
+-  div: [Function (anonymous)],
+-  dbl: [Function (anonymous)],
+-  hashToFr: [Function (anonymous)],
+-  hashAndMapToG1: [Function (anonymous)],
+-  hashAndMapToG2: [Function (anonymous)],
+-  pow: [Function (anonymous)],
+-  pairing: [Function (anonymous)],
+-  millerLoop: [Function (anonymous)],
+-  precomputedMillerLoop: [Function (anonymous)],
+-  precomputedMillerLoop2: [Function (anonymous)],
+-  precomputedMillerLoop2mixed: [Function (anonymous)],
+-  finalExp: [Function (anonymous)]
++  ptrToAsciiStr: [Function (anonymous)],
++  toHex: [Function (anonymous)],
++  toHexStr: [Function (anonymous)],
++  asciiStrToPtr: [Function (anonymous)],
++  curveType: undefined
+ }
