--- tmp-for-ts-migration/bench/js-1
@@ -1,62 +1,0 @@
-name=BLS12_381
-ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f200152d
-precomputedMillerLoop2 1735.028 usec
-precomputedMillerLoop2mixed 3595.96 usec
-all ok
-Capi benchmark
-Fr
-Fr::add 0.054 usec
-Fr::sub 0.028 usec
-Fr::mul 0.101 usec
-Fr::sqr 0.103 usec
-Fr::div 15.103 usec
-Fp
-Fp::add 0.045 usec
-Fp::sub 0.036 usec
-Fp::mul 0.269 usec
-Fp::sqr 0.27 usec
-Fp::div 27.352 usec
-Fp2
-Fp2::add 0.057 usec
-Fp2::sub 0.047 usec
-Fp2::mul 0.88 usec
-Fp2::sqr 0.577 usec
-Fp2::div 26.489 usec
-benchmark
-Fr::setByCSPRNG 16.037 usec
-pairing 5314.889 usec
-millerLoop 2344.814 usec
-finalExp 3009.967 usec
-precomputedMillerLoop 1756.858 usec
-G1::add 5.693 usec
-G1::dbl 3.54 usec
-G1::mul 461.184 usec
-G2::add 14.514 usec
-G2::dbl 6.631 usec
-G2::mul 1101.234 usec
-hashAndMapToG1 669.825 usec
-hashAndMapToG2 1408.648 usec
-G1::isValidOrder 364.879 usec
-G2::isValidOrder 475.557 usec
-Fp
-Fp::add 0.662 usec
-Fp::sub 0.641 usec
-Fp::mul 0.918 usec
-Fp::sqr 0.883 usec
-Fp::inv 26.805 usec
-Fp2
-Fp2::add 1.653 usec
-Fp2::sub 1.603 usec
-Fp2::mul 2.296 usec
-Fp2::sqr 2.2 usec
-Fp2::inv 28.786 usec
-Fr
-Fr::add 0.651 usec
-Fr::sub 0.818 usec
-Fr::mul 0.681 usec
-Fr::sqr 0.85 usec
-Fr::inv 15.357 usec
-GT::add 1.979 usec
-GT::mul 19.535 usec
-GT::sqr 13.539 usec
-GT::inv 2.016 usec
