+++ tmp-for-ts-migration/dump-specs.sh
@@ -2,9 +2,15 @@
 
 cd $(dirname ${0})
 
-node -e "mcl = require('../src'); mcl.init().then(() => console.log(mcl))" > exported-fields.orig.dump
-node -e "mcl = require('../src'); mcl.init().then(() => console.log(Object.keys(mcl).sort()))" > exported-field-names.orig.dump
+CASE_NAME=${1}
+mkdir -p ${CASE_NAME}/bench
+
+node -e "mcl = require('../src'); mcl.init().then(() => console.log(mcl))" > ${CASE_NAME}/exported-fields.dump
+node -e "mcl = require('../src'); mcl.init().then(() => console.log(Object.keys(mcl).sort()))" > ${CASE_NAME}/exported-field-names.dump
+
+diff -u {00-original,${CASE_NAME}}/exported-fields.dump > ${CASE_NAME}/exported-fields.diff
+diff -u {00-original,${CASE_NAME}}/exported-field-names.dump > ${CASE_NAME}/exported-field-names.diff
 
 for i in $(seq 1 3); do
-  node ../test/bench.js > bench/js-${i}
+  node ../test/bench.js > ${CASE_NAME}/bench/${i}
 done
